[io-layer] Improve waiting on processes a lot
We now install a SIGCHLD handler when creating a process, and we do the proper
waitpid in the SIGCHLD handler so that the process isn't left as a zombie. We
use a new data structure, MonoProcess, to keep track of created processes, and
where we have a semaphore the SIGCHLD handler releases when the process exits.
This enables us to wait properly (no polling!) for child processes, and
process_wait has been completely reworked to take advantage of this.

The ability to wait for non-child processes is partly there (when shared
handles are enabled any process can wait for a process any other mono process
has created by polling on the pid).
(cherry picked from commit afb1937e56100e368bc339045a685ef3c3b58e81)