Array's internal quick sort is no longer greedy when moving the left "wall".
Large arrays with lots of duplicate values can result in stack overflow when the left "wall" is moved in when the value is less than or equal to the pivot. Restoring the previous behavior where it was only moved if the value was less than the pivot. This results in more swaps, but similar performance; except in the case where there are lots of duplicate values, which is considerably faster when not greedily moving the left "wall" in.

Updated the algorithm to only recurse on the smaller partition, avoiding the stack overhead when recursing on the larger partition.