Making the behavior of string.CompareOrdinal(string, int, string, int, int) and string.CompareOrdinalCaseInsensitive(string, int, string, int, int) match Microsoft's implementation.
The MSDN documentation states "Because CompareOrdinal is a static method, strA and strB can be null. If both values are null, the method returns 0 (zero), which indicates that strA and strB are equal. If only one of the values is null, the method considers the non-null value to be greater." Previously, CompareOrdinal and CompareOrdinalCaseInsensitive would throw a NullReferenceException if strA or strB was null.