Increase handle collection expired interval and force update before collection.
This should prevent a race condition in which file share information is
wrongly deleted after a long sleep.

Previously, if the app slept for more than 1 minute and after waking up
the _wapi_handle_collect() method ran before the
_wapi_handle_update_refs() method, file share information could get
wiped out, which would lead to a crash when the file handle was closed.

The expiration interval is bumped to 1 day to prevent the most likely
 scenario where a user puts his machine to sleep for the night and
crashes when reopening his machine in the morning.