Exclude the current thread when stopping the world if Instruments is attached
When Instruments is attached to profile memory usage, it injects liboainject.dylib into the target process. This replaces the default mmap/munmap with version that lock for statistics collection. This will result in deadlock within Mono when the GC stops the world to collect and attemps to mmap. There is an internal __OAExcludeMachThreadID method which adds/removes the passed in thread from the internal exclusion list and prevents Instruments from collecting statistics for that thread. We add the current thread to the exclusion list after we stop the world, then remove the current thread from the exclusion list when we resume.